qcom: rpm-smd: Use arch_init and implement deferred probe mechanism
Driver init calls are serialized using a BoardFile on current targets.
This will not be the case with newer targets. Use arch_init to initialize
rpm-smd driver early in the boot process.
Not having a BoardFile to serialize the driver init calls, leaves open the
possibility of another driver being initialized and calling into rpm-smd
driver before the rpm-smd driver is initialized. Provide a mechanism to
defer calls into rpm-smd until it is probed. Return -EPROBE_DEFER if any
of the rpm-smd APIs get called before it is probed.

Change-Id: I3cc9581a4df61ba3216c20c8fab1271603e2608b
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>