iommu: msm: Remove unneeded code
The V7S mapping code is checking whether a new mapping will overlap an
old one so that we don't try to map and then fail with a partial
mapping. However, since this functionality was added the mapping code
has been changed to unwind any mappings in case of error. Thus the
check_range function is not needed anymore.

Change-Id: I706d9e189d11a5067677fc3ee5caf4dfd6a9eb14
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>