qcom: rpm-smd: Force smd channel update in interrupt lock context
In the power collapse path, when checking for availability of an ack from
rpm, use smd_is_pkt_avail. This api is designed to be used with
interrupts disabled, and will be able to poll for acks from rpm.

Wait for upto 500us to receive this ack. If in 500us, an ack is not
received generate a kernel panic.

The rpm-smd receive buffer is able to hold a limited number of packets.
Reduce the threshold of sent packets, such that apps will process received
packets in the receive buffer after 10 sent packets. This will ensure that
the receive buffer does not overflow even if the rpm sends bigger nack
packets along with ack packets.

Change-Id: Iea8019eb130986cda032aa9707d5c1e69d736ed5
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>