msm: rpm-smd: Increase the max outstanding sleep set messages to 24
When sending sleep-set requests send up 24 requests before reading any
acks.  The RX FIFO can hold 24 acks. Increase the threshold to the
max number of acks the FIFO can hold, so as to delay the process of
emptying the FIFO.

Also check and return an err on timeout and propagate that error up the
call chain.

Change-Id: I1958461c2fa3c68093d8d9aedd1e66b18cf4fbcc
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>