Bluetooth: Release locks before sleeping for L2CAP socket shutdown
If there are unacknowledged frames during an ongoing L2CAP transfer and
if socket is shutdown, it will result in deadlock as locks are not
released before going to sleep.

The L2CAP receive thread will be waiting on locks to update unacknowledged
frame count, whereas socket shutdown thread will be sleeping till
unacknowledged count becomes 0 resulting in deadlock.

Change-Id: I20d361b46a6aff02a45b4260b468f41c7ff078c2
Signed-off-by: Rupesh Tatiya <rtatiya@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>